Pin the environment, version the results, fix the README
Makes the repository reproducible from a clone alone, with no access to the
machine the campaign ran on.

Environment: requirements.txt records the exact versions of the venv that
produced everything under runs/ (Python 3.13.5, torch 2.13.0+cu130, numpy
2.5.1, pandas 3.0.3, h5py 3.16.0, matplotlib 3.11.0, PyYAML 6.0.3) -- the
only six runtime dependencies. pyproject.toml keeps them unpinned so
`pip install -e .` still resolves anywhere.

runs/ moves from fully ignored to a whitelist: figures (*.png), metrics
(*.json/*.csv/*.txt), the per-run config.yaml/config.resolved.yaml, logs and
checkpoints (*.pth) are now versioned -- 151 MB total. The regenerable *.npz
prediction arrays (5.3 GB) and *.gif animations (400 MB) stay out. Since the
checkpoints and metrics.json are in, every analysis/*.py runs straight after
cloning, with no training and no dataset.

README fixes, each of which blocked reproduction:
- Quickstart generated beam_dataset_simple/complex.h5 while registry.py
  looks up beam_dataset_A/B/C/D.h5 -- added the required --output names and
  documented the four datasets plus the three derived variants
- it still described runs as `configs/runs/<run_id>.yaml`, a layout removed
  in eb5aced; a run is now a self-contained runs/<phase>/<run_id>/config.yaml
  whose identity is its folder name, with run_id rejected inside the file
- dead link to a GUIDE.md that does not exist -> point at cli.py
- `dataset` field values listed as simple/complex -> A/B/C/D and variants
- layout still described archive/ + Archives/, consolidated into archives/

+++ src/beamsurrogate/training/teacher_forcing.py
@@ -58,7 +58,7 @@ def run(model, train_loader, X_val, y_val, cfg, model_path: Path, patience: int 
         model.eval()
         with torch.no_grad():
             pred_val = model(torch.tensor(X_val)).numpy()
-        val_loss = ((pred_val - y_val) ** 2).mean()
+        val_loss = float(((pred_val - y_val) ** 2).mean())
         scheduler.step(val_loss)
 
         train_history.append(train_loss)
